# VMS-9437: Review fix.
Graft VMS-9342 from vms
VMS-10005: Mac OS X. 'Invalid license' dialog appears only after 'System Administration' closing
VMS-10088: Added dependency from AGC mode
VMS-10089: Add EULA checkbox to the windows installer
VMS-10113: New camera settings - Motion tab fixes
VMS-10116: added video encoders tokens mapping for Uniview encoders\
VMS-10117: Build NxTool installer as an exe-file
VMS-10125: P2p message bus potential crash fixed
VMS-10259: Clien has crashed on assert on cloud settings page dialog
VMS-10313: MAC detection in autodiscovery mmode made disabled for DWC-M?T4Wi* cameras
VMS-10325: Support Wearable Cameras in General Tab of new Camera Settings dialog
VMS-7457: review fix
VMS-7661: New camera settings - crop stream URL fields if they're too long
VMS-7661: New camera settings - increased spacing between Vendor, Model & IP columns
VMS-8140 Do not ISAPI if it is not enabled
VMS-8140 Do not ISAPI if it is not enabled.
VMS-8177: review fix
VMS-8238: Externally added update file related changes.
VMS-8238: Fix && refactor Updates2Manager unit tests.
VMS-8238: Fix build.
VMS-8238: Further updates related fixes & new code.
VMS-8238: Implement RemovedData serialization & deserialization and relevant UT.
VMS-8238: Implement onFileAdded downloader signal processing and corresponding unit tests
VMS-8270: Temporary disable filtering for lens control parameters
VMS-8374: Take icon from the right path
VMS-8374: Use RESOURCE_ALIAS_PREFIX instead of copying favicon into the build directory
VMS-8494: OSC (object removal) event disabled
VMS-8494: bug fix in dw_mtt/manifest.json
VMS-8494: camera white list extended, face detection event disbled
VMS-8675: Removed 'Event Log' 'Event Rules' and 'Filter buttons' from right panel
VMS-8676: Action text corrected
VMS-8676: Implemented context menu for notification panel
VMS-8676: Replaced 'Toggle Notifications' action with 'Pin Notifications' action
VMS-8677: CR fixes
VMS-8677: CR fixes 2
VMS-8677: Implemented 'No new modifications' in right notification bar
VMS-8682: Notification tile is kept hovered when entering close button (functionality removed)
VMS-8683: Added 'CloseButton' widget for Event Tiles
VMS-9242: Fix editing accessible resources for the custom user
VMS-9307 Fixing HintButton behaviour
VMS-9475: Rename 'Storage failure' to 'Storage issue' in the event log/rules (backport)
VMS-94: Disable quality/fps selection for rtsp links
VMS-9692: Load testing results are not good in 3.2
VMS-9743: hik 2DE7330IW ptz mapping
VMS-9810: hik camera digital IO fixed
VMS-9816 Fix Client must pull video from interfaces according to 'Routing Management' table
VMS-9847: Server fails to grab video from Hikvision DS-2DF6223-CX
VMS-9984: Fix DW input trigger
VMS-9999: Implement parameter value based advanced tab controls filtering
VMS-9999: Removed attribute check for 'Initialize Focus' button

<==== Mercurial data ====>
Branch: master

+++ common_libs/nx_update/src/nx/update/manager/detail/updates2_manager_base.cpp
@@ -8,6 +8,7 @@
 #include <nx/fusion/model_functions.h>
 #include <nx/utils/log/log.h>
 #include <nx/utils/scope_guard.h>
+#include <nx/utils/literal.h>
 #include <api/global_settings.h>
 #include <nx/vms/common/p2p/downloader/downloader.h>
 
@@ -19,8 +20,7 @@ namespace detail {
 
 using namespace vms::common::p2p::downloader;
 
-Updates2ManagerBase::Updates2ManagerBase():
-    m_updateRegistry(info::UpdateRegistryFactory::create())
+Updates2ManagerBase::Updates2ManagerBase()
 {
 }
 
@@ -28,6 +28,7 @@ void Updates2ManagerBase::atServerStart()
 {
     using namespace std::placeholders;
 
+    m_updateRegistry = info::UpdateRegistryFactory::create(peerId());
     loadStatusFromFile();
     checkForGlobalDictionaryUpdate();
 
@@ -58,7 +59,7 @@ void Updates2ManagerBase::atServerStart()
     connectToSignals();
 
     m_timerManager.addNonStopTimer(
-        std::bind(&Updates2ManagerBase::checkForRemoteUpdate, this, _1),
+        std::bind(&Updates2ManagerBase::checkForRemoteUpdate, this, _1, false),
         std::chrono::milliseconds(refreshTimeout()),
         std::chrono::milliseconds(0));
 }
@@ -75,13 +76,18 @@ void Updates2ManagerBase::checkForGlobalDictionaryUpdate()
     if (!globalRegistry)
         return;
 
-    updateRegistryIfNeeded(std::move(globalRegistry));
+    {
+        QnMutexLocker lock(&m_mutex);
+        m_updateRegistry->merge(globalRegistry.get());
+    }
+
     refreshStatusAfterCheck();
 }
 
-void Updates2ManagerBase::checkForRemoteUpdate(utils::TimerId /*timerId*/)
+void Updates2ManagerBase::checkForRemoteUpdate(utils::TimerId /*timerId*/, bool forced)
 {
     auto onExitGuard = makeScopeGuard([this]() { remoteUpdateCompleted(); });
+    if (!forced)
     {
         QnMutexLocker lock(&m_mutex);
         switch (m_currentStatus.state)
@@ -101,36 +107,22 @@ void Updates2ManagerBase::checkForRemoteUpdate(utils::TimerId /*timerId*/)
     }
 
     auto remoteRegistry = getRemoteRegistry();
-    if (remoteRegistry)
+    auto globalRegistry = getGlobalRegistry();
+    QByteArray serializedUpdatedRegistry;
     {
-        updateRegistryIfNeeded(std::move(remoteRegistry));
-
-        auto globalRegistry = getGlobalRegistry();
-        QByteArray serializedUpdatedRegistry;
+        QnMutexLocker lock(&m_mutex);
+        m_updateRegistry->merge(remoteRegistry.get());
+        if (!globalRegistry || !m_updateRegistry->equals(globalRegistry.get()))
         {
-            QnMutexLocker lock(&m_mutex);
-            if (!globalRegistry || !m_updateRegistry->equals(globalRegistry.get()))
-            {
-                m_updateRegistry->merge(globalRegistry.get());
-                serializedUpdatedRegistry = m_updateRegistry->toByteArray();
-            }
+            m_updateRegistry->merge(globalRegistry.get());
+            serializedUpdatedRegistry = m_updateRegistry->toByteArray();
         }
-
-        if (!serializedUpdatedRegistry.isNull())
-            updateGlobalRegistry(serializedUpdatedRegistry);
     }
+
+    if (!serializedUpdatedRegistry.isNull())
+        updateGlobalRegistry(serializedUpdatedRegistry);
 
     refreshStatusAfterCheck();
-}
-
-void Updates2ManagerBase::updateRegistryIfNeeded(update::info::AbstractUpdateRegistryPtr other)
-{
-    QnMutexLocker lock(&m_mutex);
-    if (!other->equals(m_updateRegistry.get()))
-    {
-        other->merge(m_updateRegistry.get());
-        m_updateRegistry = std::move(other);
-    }
 }
 
 void Updates2ManagerBase::refreshStatusAfterCheck()
@@ -199,18 +191,27 @@ api::Updates2StatusData Updates2ManagerBase::download()
 
     nx::update::info::FileData fileData;
     const auto result = m_updateRegistry->findUpdateFile(
-        detail::UpdateFileRequestDataFactory::create(isClient()), &fileData);
+        detail::UpdateFileRequestDataFactory::create(isClient()),
+        &fileData);
     NX_ASSERT(result == update::info::ResultCode::ok);
 
     if (result != update::info::ResultCode::ok)
     {
         setStatus(
-            api::Updates2StatusData::StatusCode::error, "Failed to get update file information");
+            api::Updates2StatusData::StatusCode::error,
+            "Failed to get update file information");
         return m_currentStatus.base();
     }
 
     using namespace vms::common::p2p::downloader;
-    NX_ASSERT(!fileData.file.isNull());
+    NX_ASSERT(!fileData.isNull());
+    if (fileData.isNull())
+    {
+        setStatus(
+            api::Updates2StatusData::StatusCode::error,
+            "File data is empty");
+        return m_currentStatus.base();
+    }
 
     auto fileInformation = downloader()->fileInformation(fileData.file);
     bool alreadyHasFileInDownLoader = fileInformation.isValid();
@@ -218,12 +219,18 @@ api::Updates2StatusData Updates2ManagerBase::download()
     {
         NX_ASSERT(!fileData.md5.isNull());
         NX_ASSERT(fileData.size != 0);
+        if (fileData.md5.isNull() || fileData.size == 0)
+        {
+            setStatus(
+                api::Updates2StatusData::StatusCode::error,
+                "File data is corrupted");
+            return m_currentStatus.base();
+        }
 
         fileInformation.md5 = QByteArray::fromHex(fileData.md5.toBase64());
         fileInformation.name = fileData.file;
         fileInformation.size = fileData.size;
         fileInformation.url = fileData.url;
-        fileInformation.peerPolicy = FileInformation::PeerSelectionPolicy::byPlatform;
 
         downloader()->deleteFile(fileData.file);
         if (!m_currentStatus.file.isEmpty())
@@ -231,6 +238,8 @@ api::Updates2StatusData Updates2ManagerBase::download()
 
         m_currentStatus.file.clear();
     }
+    fileInformation.peerPolicy = FileInformation::PeerSelectionPolicy::byPlatform;
+    fileInformation.additionalPeers = m_updateRegistry->additionalPeers(fileInformation.name);
 
     ResultCode resultCode = downloader()->addFile(fileInformation);
     switch (resultCode)
@@ -254,8 +263,7 @@ api::Updates2StatusData Updates2ManagerBase::download()
                             QString("Update file is corrupted: %1").arg(fileInformation.name));
                         downloader()->deleteFile(fileInformation.name, /*deleteData*/ true);
                         m_currentStatus.file.clear();
-                        // #TODO #akulikov deal with below
-//                        m_updateRegistry->removeManualFileData(fileInformation.name);
+                        m_updateRegistry->removeFileData(fileInformation.name);
                         break;
                     case FileInformation::Status::downloading:
                         setStatus(
@@ -330,6 +338,10 @@ void Updates2ManagerBase::startPreparing(const QString& updateFilePath)
         updateFilePath,
         [this](PrepareResult prepareResult)
         {
+            QnMutexLocker lock(&m_mutex);
+            if (m_currentStatus.state != api::Updates2StatusData::StatusCode::preparing)
+                return;
+
             switch (prepareResult)
             {
                 case PrepareResult::ok:
@@ -445,31 +457,45 @@ void Updates2ManagerBase::onFileAdded(const FileInformation& fileInformation)
     if (manualFileData.isNull())
         return;
 
-    QByteArray serializedGlobalRegistry;
+    if (fileInformation.status == FileInformation::Status::notFound
+        || fileInformation.status == FileInformation::Status::corrupted)
+    {
+        NX_VERBOSE(
+            this,
+            lm("External update file %1 added, but has an inappropriate status")
+                .arg(fileInformation.name));
+        return;
+    }
+
     {
         QnMutexLocker lock(&m_mutex);
-        if (m_currentStatus.file == fileInformation.name)
-            return;
-
-        NX_VERBOSE(this, lm("External update file %1 added to the Downloader")
-            .args(fileInformation.name));
+        NX_VERBOSE(
+            this,
+            lm("External update file %1 was added to the Downloader").args(fileInformation.name));
 
         manualFileData.peers.append(peerId());
         m_updateRegistry->addFileData(manualFileData);
-        auto globalRegistry = getGlobalRegistry();
-        if (!m_updateRegistry->equals(globalRegistry.get()))
-        {
-            m_updateRegistry->merge(globalRegistry.get());
-            serializedGlobalRegistry = m_updateRegistry->toByteArray();
-        }
     }
 
-    updateGlobalRegistry(serializedGlobalRegistry);
+    checkForRemoteUpdate(utils::TimerId(), /* forced */ true);
 }
 
-void Updates2ManagerBase::onFileDeleted(const QString& /*fileName*/)
+void Updates2ManagerBase::onFileDeleted(const QString& fileName)
 {
-    // #TODO #akulikov Implement this.
+    auto manualFileData = update::info::ManualFileData::fromFileName(fileName);
+    if (manualFileData.isNull())
+        return;
+
+    {
+        QnMutexLocker lock(&m_mutex);
+        NX_VERBOSE(
+            this,
+            lm("External update file %1 was removed from the Downloader").args(fileName));
+
+        m_updateRegistry->removeFileData(fileName);
+    }
+
+    checkForRemoteUpdate(utils::TimerId(), /* forced */ true);
 }
 
 void Updates2ManagerBase::onFileInformationChanged(const FileInformation& /*fileInformation*/)
@@ -512,6 +538,41 @@ api::Updates2StatusData Updates2ManagerBase::install()
         setStatus(api::Updates2StatusData::StatusCode::installing, "Installing update");
     else
         setStatus(api::Updates2StatusData::StatusCode::error, "Error while installing update");
+
+    return m_currentStatus.base();
+}
+
+api::Updates2StatusData Updates2ManagerBase::cancel()
+{
+    QnMutexLocker lock(&m_mutex);
+    switch (m_currentStatus.state)
+    {
+    case api::Updates2StatusData::StatusCode::notAvailable:
+        break;
+    case api::Updates2StatusData::StatusCode::available:
+    case api::Updates2StatusData::StatusCode::error:
+    case api::Updates2StatusData::StatusCode::checking:
+        setStatus(
+            api::Updates2StatusData::StatusCode::notAvailable,
+            "Update is unavailable");
+        break;
+    case api::Updates2StatusData::StatusCode::downloading:
+        NX_ASSERT(!m_currentStatus.file.isEmpty());
+        downloader()->deleteFile(m_currentStatus.file);
+        m_updateRegistry->removeFileData(m_currentStatus.file);
+        setStatus(
+            api::Updates2StatusData::StatusCode::notAvailable,
+            "Update is unavailable");
+        break;
+    case api::Updates2StatusData::StatusCode::readyToInstall:
+    case api::Updates2StatusData::StatusCode::preparing:
+        setStatus(
+            api::Updates2StatusData::StatusCode::available,
+            "Update is unavailable");
+        break;
+    case api::Updates2StatusData::StatusCode::installing:
+        break;
+    }
 
     return m_currentStatus.base();
 }
